backend/requirements.txt
flask
flask-cors
pandas
numpy
python-dotenv
google-generativeai